[NS-72] Move metrics and serializer code to dedicated  package
- Move all metrics-related code to `io.sparkcyclone.metrics` package

- Move all serializer-related code over to `io.sparkcyclone.data.serializer` package
