+++ src/main/scala/io/sparkcyclone/cache/ArrowBasedCacheSerializer.scala
@@ -3,7 +3,7 @@ package io.sparkcyclone.cache
 import io.sparkcyclone.colvector.ArrowVectorConversions._
 import io.sparkcyclone.colvector.SparkSqlColumnVectorConversions._
 import io.sparkcyclone.util.CallContext
-import io.sparkcyclone.ve.VeProcessMetrics
+import io.sparkcyclone.metrics.VeProcessMetrics
 import org.apache.arrow.memory.BufferAllocator
 import org.apache.arrow.vector.types.pojo.Schema
 import org.apache.spark.TaskContext
