+++ src/main/scala/io/sparkcyclone/colvector/VeColVector.scala
@@ -2,7 +2,7 @@ package io.sparkcyclone.colvector
 
 import io.sparkcyclone.cache.VeColColumnarVector
 import io.sparkcyclone.spark.agile.core._
-import io.sparkcyclone.ve.VeProcessMetrics
+import io.sparkcyclone.metrics.VeProcessMetrics
 import io.sparkcyclone.vectorengine.{VeAsyncResult, VeProcess}
 import org.apache.spark.sql.vectorized.ColumnVector
 import org.bytedeco.javacpp.BytePointer
